Chart finder and open new file fix for 1.03
There was a bug where any empty white space at the end of a file would register as a new file; this was fixed. Also fixed a small issue with the index not being reset to 0 for the chart chooser. This caused an unavailable chart to still be selected when in a new file.