Added probability function to Index
I need to tweak the numbers a bit, but this attempts to add a bit more control to the random placement of notes, so notes don't keep getting placed in the same lane as much. This should also be applicable to random note placement for spread, but I need to take a futher look at that later.