Added probability function to the solo/kb7 converter & attempted to added another handler.
Probability for placement includes the previous requirements that existed before, but now the initial random placement is handled entirely by how likely a note should be in a specific spot. I also attempted to catch another one of the cases that appears, where there are a bunch of hold notes above and below, as well as a bunch of regular notes above. I have an idea of how to handle this much better, but that'll be for later.